backend/requirements_txt.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.20
requests>=2.31.0
jwt>=1.3.1
pyjwt>=2.8.0
python-multipart>=0.0.6
openai>=1.3.7
google-generativeai>=0.3.1
typing_extensions>=4.8.0
bcrypt>=4.0.1
